feat: allow preserving tomcat base directory on server stop (#6)
* feat: allow preserving tomcat base directory on server stop

Keeping content of the base directory can sometimes be handy,
for example when third party application packaged as war is required
by the server under test. In my case it is gitblit which is unpacked
by maven in generate test resources phase and then gitblit is started
via junit-servers for my application integration tests.
It is much more efficient to preserve tomcat directory between restarts
(even when some small housekeeping is necessary) than copying over
megabytes of jars to temporary directory between gitblit restarts.

+++ junit-servers-tomcat/src/main/java/com/github/mjeanroy/junit/servers/tomcat/EmbeddedTomcat.java
@@ -220,7 +220,9 @@ public class EmbeddedTomcat extends AbstractEmbeddedServer<Tomcat, EmbeddedTomca
 				context = null;
 			}
 
-			deleteDirectory(configuration.getBaseDir());
+			if (!configuration.isKeepBaseDir()) {
+				deleteDirectory(configuration.getBaseDir());
+			}
 		}
 		catch (Exception ex) {
 			throw new ServerStopException(ex);
